Move user endpoint logic to user controller file
