+++ server/routes/user-routes.js
@@ -1,0 +1,9 @@
+const express = require('express');
+const router = express.Router();
+
+const userController = require('../controller/user-controller');
+
+router.route("/register")
+    .post(userController.createUser);
+
+module.exports = router;
